improve performance of _determineActualTypes
The current algorithm is inefficient for unary and binary types. Given
an array containing 1000 elements, for example, we first determine the
types of each element, resulting in an array of 1000 arrays of types.
We then find the intersection of these "sets". This commit introduces
a different approach: for each element, we refine the set of types of
which all previous elements are members rather than filtering the whole
environment each time. This is significantly more efficient for large
arrays.